requirements.txt
click==8.1.7
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.12.0
anthropic==0.18.1
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
fastapi==0.110.0
uvicorn==0.27.1
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.6.3
python-multipart==0.0.9
jinja2==3.1.3
redis==5.0.3
langchain==0.1.11
langchain-community==0.0.25
transformers==4.38.2
torch==2.2.1
datasets==2.17.1
pytest==8.1.1
pytest-asyncio==0.23.5